requirements.txt
Pillow
mediapipe
customtkinter
numpy
opencv-python
rembg
onnxruntime
